Add tab configurations and localization support for user and transporter roles
- Introduce TRANSPORTER_TABS and USER_TABS constants to define tab configurations for different user roles, enhancing navigation organization.
- Create a new types/navigation.ts file to define the TabConfig interface and IconName type for better type safety.
- Update useTabConfigurations hook to utilize the new tab configurations, streamlining tab management.
- Add comments to hooks for improved code clarity and maintainability.
