+++ hooks/useTokenInitialization.ts
@@ -2,6 +2,7 @@ import { useEffect } from "react";
 import { useStore } from "@/store/useStore";
 import TokenManager from "@/features/auth/services/token-manager";
 
+// Hook to initialize the token manager
 export const useTokenInitialization = () => {
   const { accessToken } = useStore();
