+++ types/navigation.ts
@@ -1,0 +1,9 @@
+import { Ionicons } from "@expo/vector-icons";
+
+export type IconName = keyof typeof Ionicons.glyphMap;
+
+export interface TabConfig {
+  name: string;
+  titleKey: string;
+  icon: IconName;
+}
